Fix up proof for shift widths change
Verilog shifts don't need the data and shift amount widths to
match. The shift amount is just taken for its integer
value. However, in SMTLIB, the `bv_shr` and `bv_shl` operators expect
their left- and right-hand sides to have the same width.

Therefore, to compile a verilog shift, I need to cast both lhs and rhs
up to their max width, and then select down to the width of the
rhs. This is exactly equivalent to just doing the shift without the
extension/selection, which is what is being proved in this commit.

They key insight for the proof is that shifts can be represented as
selection (keeping what hasn't fallen off from the shift) followed by
a concat with the zeros that have been shifted in. Using this rewrite,
our top level proof becomes one on only `extr`s and `concat`s. Add in
some lemmas to manipulate such combinations (pushing `extr`s into
`concat`s and such), and you can essentially prove the whole thing
using only rewrites.

This commit also adds tactics for working with bitvectors and related things:
- `bitvector_erase` takes a goal on sized bitvectors to one on raw
bitvectors
- `N_to_nat` eliminates all `N`s in the hyps and goal and replaces
them with `nat`s
- `destruct_mins` destructs cases for each `min`
- `kill_bools` eliminates replaces `(_ <=? _) = ?b` assumptions with
(_ <= _) ones.
- The `xbv_size` rewrite db tries to push calls to `XBV.size` as far
down as possible.